latest

+++ scripts/b3/eval.sh
@@ -17,10 +17,12 @@ do
     #echo $model
     save_path="$(dirname $i)/eval.pt"
 
-    last=$(tail -c2 $save_path | head -c1)
-    if [ "$last" = "=" ]; then
-        echo "bad error with $save_path"
-        rm $save_path
+    if [ -f "$save_path" ]; then
+        last=$(tail -c2 $save_path | head -c1)
+        if [ "$last" = "=" ]; then
+            echo "bad error with $save_path"
+            rm $save_path
+        fi
     fi
 
     if [ -f "$save_path" ]; then
@@ -29,18 +31,36 @@ do
         # rm "$(dirname $i)/epoch_3.pt"
         # rm "$(dirname $i)/epoch_4.pt"
         # rm "$(dirname $i)/epoch_latest.pt"
+        # if [[ $save_path == *"int8"* ]]; then
+        #     rm $save_path
+        #     echo $save_path
+        # fi
         echo "$save_path exists."
     else
-        if [[ $save_path == *"extraln"* ]]; then
+        if [[ $save_path == *"int8"* ]]; then
+            echo "$save_path does not exist -- int8."
+            torchrun --nproc_per_node 2 -m training.main \
+            --batch-size 200   --workers 4 --model $model  --train-num-samples 413000000  \
+            --local-loss  --gather-with-grad     --grad-checkpointing       --precision custom_fp16  \
+            --save-most-recent --pretrained $i --custom-attention vanilla --int8mix \
+            --imagenet-val /fsx/rom1504/imagenetval/imagenet_validation &> $save_path
+        elif [[ $save_path == *"fp8"* ]]; then
+            echo "$save_path does not exist -- fp8."
+            torchrun --nproc_per_node 2 -m training.main \
+            --batch-size 200   --workers 4 --model $model  --train-num-samples 413000000  \
+            --local-loss  --gather-with-grad     --grad-checkpointing       --precision custom_fp16  \
+            --save-most-recent --pretrained $i --custom-attention vanilla --fp8mix \
+            --imagenet-val /fsx/rom1504/imagenetval/imagenet_validation &> $save_path
+        elif [[ $save_path == *"extraln"* ]]; then
             echo "$save_path does not exist -- extraln."
-            torchrun --nproc_per_node 4 -m training.main \
+            torchrun --nproc_per_node 2 -m training.main \
             --batch-size 200   --workers 4 --model $model  --train-num-samples 413000000  \
             --local-loss  --gather-with-grad     --grad-checkpointing       --precision amp_bfloat16  \
             --save-most-recent --pretrained $i --custom-attention extra_ln \
             --imagenet-val /fsx/rom1504/imagenetval/imagenet_validation &> $save_path
         else
             echo "$save_path does not exist."
-            torchrun --nproc_per_node 4 -m training.main \
+            torchrun --nproc_per_node 2 -m training.main \
             --batch-size 200   --workers 4 --model $model  --train-num-samples 413000000  \
             --local-loss  --gather-with-grad     --grad-checkpointing       --precision amp_bfloat16  \
             --save-most-recent --pretrained $i --custom-attention vanilla \
@@ -49,12 +69,16 @@ do
     fi
 done
 
+# """
+# torchrun --nproc_per_node 2 -m training.main \
+# --batch-size 200   --workers 4 --model ViT-L-14  --train-num-samples 413000000  \
+# --local-loss  --gather-with-grad     --grad-checkpointing  --precision custom_fp16 --int8mix \
+# --save-most-recent \
+# --pretrained /fsx/home-mitchellw/experimetns/opt3/clipadamw-int8mix-ViT-L-14-16384-2e-3-0.98-v0/checkpoints/epoch_5.pt \
+# --custom-attention vanilla \
+# --imagenet-val /fsx/rom1504/imagenetval/imagenet_validation
 
-"""
 
-torchrun --nproc_per_node 4 -m training.main \
---batch-size 200   --workers 4 --model ViT-B-32  --train-num-samples 413000000  \
---local-loss  --gather-with-grad     --grad-checkpointing       --precision amp_bfloat16  \
---save-most-recent --pretrained /fsx/home-mitchellw/experimetns/opt3/ldog-ViT-B-32-16384-1-0.98-skipcwd1en4-v0/checkpoints/ema_1_5.pt --custom-attention vanilla \
---imagenet-val /fsx/rom1504/imagenetval/imagenet_validation
-"'"
+# 54.8
+# 57.0
+# """
